--- a/pytest-framework.ini
+++ b/pytest-framework.ini
@@ -5,8 +5,6 @@ python_files=
     test_*.py
 testpaths =
     src
-markers =
-    run_in_serial
 addopts = 
     -p pytester
     -p pytest_plugins.eels_resolver