--- a/tox.ini
+++ b/tox.ini
@@ -71,8 +71,7 @@ extras =
     test
     lint # Required `gentest` for formatting tests
 commands =
-    pytest -c ./pytest-framework.ini -n auto -m "not run_in_serial"
-    pytest -c ./pytest-framework.ini -m run_in_serial
+    pytest -c ./pytest-framework.ini -n auto
 
 
 [forks]