## 🗒️ Description

### `--t8n-server-url` fill flag

Adds a new flag to t8n in order to be able to pass an already-running t8n server to run unit tests.

This allows us to use a single eels resolver server instance in all unit tests and speed up test running considerably.

### Remove `run_in_serial`

The unit test marker `run_in_serial` is removed since now all tests can run in parallel without issues.

### Improvements
#### Before
`uvx --with=tox-uv tox -e pytest` command:
- `pytest -c ./pytest-framework.ini -n auto -m 'not run_in_serial'` - 103.30s (0:01:43)
- `pytest -c ./pytest-framework.ini -m run_in_serial` - 96.13s (0:01:36)
-  Total - 199.43s (0:03:19)
#### After
`uvx --with=tox-uv tox -e pytest` command:
- `pytest -c ./pytest-framework.ini -n auto` - 29.46s (0:00:29)
-  Total - 29.46s (0:00:29)

## 🔗 Related Issues or PRs
N/A.

## ✅ Checklist
- [ ] All: Ran fast `tox` checks to avoid unnecessary CI fails, see also [Code Standards](https://eest.ethereum.org/main/getting_started/code_standards/) and [Enabling Pre-commit Checks](https://eest.ethereum.org/main/dev/precommit/):
    ```console
    uvx --with=tox-uv tox -e lint,typecheck,spellcheck,markdownlint
    ```
- [ ] All: PR title adheres to the [repo standard](https://eest.ethereum.org/main/getting_started/contributing/?h=contri#commit-messages-issue-and-pr-titles) - it will be used as the squash commit message and should start `type(scope):`.
- [ ] All: Considered adding an entry to [CHANGELOG.md](/ethereum/execution-spec-tests/blob/main/docs/CHANGELOG.md).
- [ ] All: Considered updating the online docs in the [./docs/](/ethereum/execution-spec-tests/blob/main/docs/) directory.
- [ ] All: Set appropriate labels for the changes (only maintainers can apply labels).
- [ ] Tests: Ran `mkdocs serve` locally and verified the auto-generated docs for new tests in the [Test Case Reference](https://eest.ethereum.org/main/tests/) are correctly formatted.
- [ ] Tests: For PRs implementing a missed test case, update the [post-mortem document](/ethereum/execution-spec-tests/blob/main/docs/writing_tests/post_mortems.md) to add an entry the list.
- [ ] Ported Tests: All converted JSON/YML tests from [ethereum/tests](/ethereum/tests) or [tests/static](/ethereum/execution-spec-tests/blob/main/tests/static) have been assigned `@ported_from` marker.